chore(end-to-end): split `tests/utils` package - `tests/utils/rpc` package - `tests/utils/websocket` package - `tests/utils/config` package
