+++ .gitignore
@@ -23,8 +23,6 @@ test_data
 trie_putandget_failed_test_data_*
 tmp
 
-tests/utils/config*
-
 # node_modules used by polkadot.js/api tests
 tests/polkadotjs_test/node_modules
 !tests/polkadotjs_test/test/*.wasm
