+++ tests/rpc/rpc_03-chain_test.go
@@ -12,6 +12,8 @@ import (
 	"github.com/ChainSafe/gossamer/dot/rpc/modules"
 	libutils "github.com/ChainSafe/gossamer/lib/utils"
 	"github.com/ChainSafe/gossamer/tests/utils"
+	"github.com/ChainSafe/gossamer/tests/utils/config"
+	websocketutils "github.com/ChainSafe/gossamer/tests/utils/websocket"
 	"github.com/gorilla/websocket"
 	"github.com/stretchr/testify/require"
 )
@@ -60,7 +62,7 @@ func TestChainRPC(t *testing.T) {
 
 	t.Log("starting gossamer...")
 	genesisPath := libutils.GetDevGenesisSpecPathTest(t)
-	config := utils.CreateDefaultConfig(t)
+	config := config.CreateDefault(t)
 	nodes, err := utils.InitializeAndStartNodes(t, 1, genesisPath, config)
 	require.NoError(t, err)
 
@@ -182,7 +184,7 @@ func TestChainSubscriptionRPC(t *testing.T) {
 
 	t.Log("starting gossamer...")
 	genesisPath := libutils.GetDevGenesisSpecPathTest(t)
-	config := utils.CreateDefaultConfig(t)
+	config := config.CreateDefault(t)
 	nodes, err := utils.InitializeAndStartNodesWebsocket(t, 1, genesisPath, config)
 	require.NoError(t, err)
 
@@ -231,13 +233,13 @@ func callWebsocket(t *testing.T, test *testCase) {
 			case int:
 				// check for result subscription number
 				resNum := 0
-				err = utils.DecodeWebsocket(v, &resNum)
+				err = websocketutils.Decode(v, &resNum)
 				require.NoError(t, err)
 
 			case map[string]interface{}:
 				// check result map response
 				resMap := make(map[string]interface{})
-				err = utils.DecodeWebsocket(v, &resMap)
+				err = websocketutils.Decode(v, &resMap)
 				require.NoError(t, err)
 
 				// check values in map are expected type
