+++ tests/stress/network_test.go
@@ -10,6 +10,8 @@ import (
 
 	libutils "github.com/ChainSafe/gossamer/lib/utils"
 	"github.com/ChainSafe/gossamer/tests/utils"
+	"github.com/ChainSafe/gossamer/tests/utils/config"
+	"github.com/ChainSafe/gossamer/tests/utils/rpc"
 
 	"github.com/ChainSafe/gossamer/internal/log"
 	"github.com/stretchr/testify/require"
@@ -19,7 +21,7 @@ func TestNetwork_MaxPeers(t *testing.T) {
 	numNodes := 9 // 9 block producers
 	genesisPath := libutils.GetGssmrGenesisRawPathTest(t)
 	utils.Logger.Patch(log.SetLevel(log.Info))
-	config := utils.CreateDefaultConfig(t)
+	config := config.CreateDefault(t)
 	nodes, err := utils.InitializeAndStartNodes(t, numNodes, genesisPath, config)
 	require.NoError(t, err)
 
@@ -36,7 +38,7 @@ func TestNetwork_MaxPeers(t *testing.T) {
 	for i, node := range nodes {
 		const getPeersTimeout = time.Second
 		getPeersCtx, cancel := context.WithTimeout(ctx, getPeersTimeout)
-		peers, err := utils.GetPeers(getPeersCtx, node.RPCPort)
+		peers, err := rpc.GetPeers(getPeersCtx, node.RPCPort)
 		cancel()
 
 		require.NoError(t, err)
