+++ tests/utils/gossamer_utils.go
@@ -18,10 +18,10 @@ import (
 	"time"
 
 	"github.com/ChainSafe/gossamer/dot"
-	ctoml "github.com/ChainSafe/gossamer/dot/config/toml"
 	"github.com/ChainSafe/gossamer/dot/rpc/modules"
 	"github.com/ChainSafe/gossamer/internal/log"
 	"github.com/ChainSafe/gossamer/lib/utils"
+	"github.com/ChainSafe/gossamer/tests/utils/rpc"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -246,13 +246,13 @@ var (
 func checkNodeStarted(ctx context.Context, gossamerHost string) error {
 	const method = "system_health"
 	const params = "{}"
-	respBody, err := PostRPC(ctx, gossamerHost, method, params)
+	respBody, err := rpc.Post(ctx, gossamerHost, method, params)
 	if err != nil {
 		return fmt.Errorf("cannot post RPC: %w", err)
 	}
 
 	target := new(modules.SystemHealthResponse)
-	err = DecodeRPC(respBody, target)
+	err = rpc.Decode(respBody, target)
 	if err != nil {
 		return fmt.Errorf("cannot decode RPC: %w", err)
 	}
@@ -437,114 +437,4 @@ func GenerateGenesisAuths(t *testing.T, numAuths int) (genesisPath string) {
 	require.NoError(t, err)
 
 	return genesisPath
-}
-
-func generateDefaultConfig() *ctoml.Config {
-	return &ctoml.Config{
-		Global: ctoml.GlobalConfig{
-			Name:           "Gossamer",
-			ID:             "gssmr",
-			LogLvl:         "crit",
-			MetricsAddress: "localhost:9876",
-			RetainBlocks:   256,
-			Pruning:        "archive",
-		},
-		Log: ctoml.LogConfig{
-			CoreLvl: "info",
-			SyncLvl: "info",
-		},
-		Init: ctoml.InitConfig{
-			Genesis: "./chain/gssmr/genesis.json",
-		},
-		Account: ctoml.AccountConfig{
-			Key:    "",
-			Unlock: "",
-		},
-		Core: ctoml.CoreConfig{
-			Roles:            4,
-			BabeAuthority:    true,
-			GrandpaAuthority: true,
-			GrandpaInterval:  1,
-		},
-		Network: ctoml.NetworkConfig{
-			Bootnodes:   nil,
-			ProtocolID:  "/gossamer/gssmr/0",
-			NoBootstrap: false,
-			NoMDNS:      false,
-			MinPeers:    1,
-			MaxPeers:    3,
-		},
-		RPC: ctoml.RPCConfig{
-			Enabled:  false,
-			Unsafe:   true,
-			WSUnsafe: true,
-			Host:     "localhost",
-			Modules:  []string{"system", "author", "chain", "state"},
-			WS:       false,
-		},
-	}
-}
-
-// CreateDefaultConfig generates a default config and writes
-// it to a temporary file for the current test.
-func CreateDefaultConfig(t *testing.T) (configPath string) {
-	cfg := generateDefaultConfig()
-	return writeTestTOMLConfig(t, cfg)
-}
-
-// CreateConfigLogGrandpa generates a grandpa config and writes
-// it to a temporary file for the current test.
-func CreateConfigLogGrandpa(t *testing.T) (configPath string) {
-	cfg := generateDefaultConfig()
-	cfg.Log = ctoml.LogConfig{
-		CoreLvl:           "crit",
-		NetworkLvl:        "debug",
-		RuntimeLvl:        "crit",
-		BlockProducerLvl:  "info",
-		FinalityGadgetLvl: "debug",
-	}
-	return writeTestTOMLConfig(t, cfg)
-}
-
-// CreateConfigNoBabe generates a no-babe config and writes
-// it to a temporary file for the current test.
-func CreateConfigNoBabe(t *testing.T) (configPath string) {
-	cfg := generateDefaultConfig()
-	cfg.Global.LogLvl = "info"
-	cfg.Log = ctoml.LogConfig{
-		SyncLvl:    "debug",
-		NetworkLvl: "debug",
-	}
-	cfg.Core.BabeAuthority = false
-	return writeTestTOMLConfig(t, cfg)
-}
-
-// CreateConfigNoGrandpa generates an no-grandpa config and writes
-// it to a temporary file for the current test.
-func CreateConfigNoGrandpa(t *testing.T) (configPath string) {
-	t.Helper()
-	cfg := generateDefaultConfig()
-	cfg.Core.GrandpaAuthority = false
-	cfg.Core.BABELead = true
-	cfg.Core.GrandpaInterval = 1
-	return writeTestTOMLConfig(t, cfg)
-}
-
-// CreateConfigNotAuthority generates an non-authority config and writes
-// it to a temporary file for the current test.
-func CreateConfigNotAuthority(t *testing.T) (configPath string) {
-	t.Helper()
-	cfg := generateDefaultConfig()
-	cfg.Core.Roles = 1
-	cfg.Core.BabeAuthority = false
-	cfg.Core.GrandpaAuthority = false
-	return writeTestTOMLConfig(t, cfg)
-}
-
-func writeTestTOMLConfig(t *testing.T, cfg *ctoml.Config) (configPath string) {
-	t.Helper()
-	configPath = filepath.Join(t.TempDir(), "config.toml")
-	err := dot.ExportTomlConfig(cfg, configPath)
-	require.NoError(t, err)
-	return configPath
 }
