+++ tests/utils/rpc/header.go
@@ -1,7 +1,7 @@
 // Copyright 2021 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
-package utils
+package rpc
 
 import (
 	"fmt"
